rename getEmails now that we return names too
