+++ handlers/sf-contact-merge/src/main/scala/com/gu/sf_contact_merge/getaccounts/GetIdentityAndZuoraEmailsForAccountsSteps.scala
@@ -2,7 +2,7 @@ package com.gu.sf_contact_merge.getaccounts
 
 import com.gu.salesforce.TypesForSFEffectsData.SFContactId
 import com.gu.sf_contact_merge.getaccounts.GetContacts.AccountId
-import com.gu.sf_contact_merge.getaccounts.GetEmails.{EmailAddress, FirstName, LastName}
+import com.gu.sf_contact_merge.getaccounts.GetZuoraContactDetails.{EmailAddress, FirstName, LastName}
 import com.gu.sf_contact_merge.update.UpdateSalesforceIdentityId.IdentityId
 import com.gu.util.resthttp.Types.{ClientFailableOp, ClientSuccess}
 import com.gu.util.zuora.SafeQueryBuilder.MaybeNonEmptyList
@@ -21,27 +21,27 @@ object GetIdentityAndZuoraEmailsForAccountsSteps {
 
   def apply(zuoraQuerier: ZuoraQuerier, accountIds: NonEmptyList[AccountId]): ClientFailableOp[List[IdentityAndSFContactAndEmail]] = {
 
-    val getEmails: NonEmptyList[GetEmails.ContactId] => ClientFailableOp[Map[GetEmails.ContactId, GetEmails.Record]] =
-      GetEmails(zuoraQuerier, _)
-    val getContacts: NonEmptyList[AccountId] => ClientFailableOp[Map[GetEmails.ContactId, GetContacts.IdentityAndSFContact]] =
+    val getZuoraContactDetails: NonEmptyList[GetZuoraContactDetails.ContactId] => ClientFailableOp[Map[GetZuoraContactDetails.ContactId, GetZuoraContactDetails.ZuoraContactDetails]] =
+      GetZuoraContactDetails(zuoraQuerier, _)
+    val getContacts: NonEmptyList[AccountId] => ClientFailableOp[Map[GetZuoraContactDetails.ContactId, GetContacts.IdentityAndSFContact]] =
       GetContacts(zuoraQuerier, _)
 
     for {
       identityForBillingContact <- getContacts(accountIds)
-      emailForBillingContact <- MaybeNonEmptyList(identityForBillingContact.keys.toList) match {
-        case Some(contactIds) => getEmails(contactIds)
+      contactDetailsForBillingContact <- MaybeNonEmptyList(identityForBillingContact.keys.toList) match {
+        case Some(contactIds) => getZuoraContactDetails(contactIds)
         case None => ClientSuccess(Map.empty[Any, Nothing])
       }
     } yield {
       identityForBillingContact.map {
         case (contact, account) =>
-          val contactRecord = emailForBillingContact(contact)
+          val contactDetails = contactDetailsForBillingContact(contact)
           IdentityAndSFContactAndEmail(
             account.identityId,
             account.sfContactId,
-            contactRecord.emailAddress,
-            contactRecord.firstName,
-            contactRecord.lastName
+            contactDetails.emailAddress,
+            contactDetails.firstName,
+            contactDetails.lastName
           )
       }.toList
     }
